Fix non suppression du fichier d'un cours texte

+++ src/controllers/classes/files/FileManager.php
@@ -12,6 +12,10 @@ abstract class FileManager
    */
   public static function delete($filePath)
   {
+    if (!$filePath) {
+      return true;
+    }
+
     if (self::exists($filePath)) {
       return unlink($filePath);
     }
